requirements.txt
amqp==2.6.1
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.14.56
botocore==1.17.56
cachelib==0.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
defusedxml==0.6.0
Deprecated==1.2.10
dj-database-url==0.5.0
dj-rest-auth==1.1.1
Django==3.1.1
django-allauth==0.42.0
django-appconf==1.0.4
django-celery-email==3.0.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-guardian==2.3.0
django-heroku==0.3.1
django-jazzmin==2.2.7
django-redis==4.12.1
django-ses==1.0.3
django-storages==1.10
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
docutils==0.15.2
drf-yasg==1.17.1
Faker==4.1.2
future==0.18.2
gevent==20.6.2
greenlet==0.4.16
humanize==2.6.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
keep==2.9
kombu==4.6.11
lxml==4.5.2
MarkupSafe==1.1.1
monotonic==1.5
oauthlib==3.1.0
packaging==20.4
prometheus-client==0.8.0
psycopg2==2.8.6
pycparser==2.20
PyGithub==1.53
Pygments==2.7.0
PyJWT==1.7.1
pyparsing==2.4.7
pyquery==1.4.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
terminaltables==3.1.0
text-unidecode==1.3
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
zope.event==4.4
zope.interface==5.1.0
